% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clans}
\alias{clans}
\title{clans}
\format{A SpatialPolygonsDataFrame with 305 features:
\describe{
\item{\code{ID}}{Object identifier}
\item{\code{LBL}}{Clan identifier}
\item{\code{FIP}}{Two character country code for Liberia}
\item{\code{MMT_ID}}{Three character country code for Liberia}
\item{\code{SHORT__FRM}}{Short form name for Liberia}
\item{\code{LONG_FRM}}{Long form name for Liberia}
\item{\code{ADM0}}{Administrative level 0 of Liberia (Country name)}
\item{\code{ADM1}}{Administrative level 1 of Liberia (County name)}
\item{\code{ADM2}}{Administrative level 2 of Liberia (District name)}
\item{\code{ADM3}}{Administrative level 3 of Liberia (Clan name)}
}}
\source{
Map Maker Ltd. (2007). Clans, Liberia, 2007 Map Maker, Ltd. Available
at: \url{http://purl.stanford.edu/vm535pw6013}.
}
\usage{
clans
}
\description{
Administrative Level 2 (clans) map of Liberia
}
\keyword{datasets}
