% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{counties}
\alias{counties}
\title{counties}
\format{A SpatialPolygonsDataFrame with 15 features:
\describe{
\item{\code{OBJECTID}}{Object identifier}
\item{\code{CNTRY_CODE}}{Three-character ISO country code. For Liberia it is
    LBR.}
\item{\code{WHO_CODE}}{Three-character WHO country code. For Liberia it is
    LIB.}
\item{\code{Hrparent}}{HR Parent}
\item{\code{Shape_Leng}}{Length of shape}
\item{\code{Shape_Area}}{Area of shape}
\item{\code{admin1name}}{Name of county}
\item{\code{admin1Pcod}}{County postal code}
\item{\code{admin0name}}{Name of country}
\item{\code{admin0Pcod}}{Country postal code. LR for Liberia.}
\item{\code{Pcod1_iso3}}{County postal code using ISO3 code}
}}
\source{
\url{https://data.humdata.org/dataset/liberia-admin-level-2-boundaries}
}
\usage{
counties
}
\description{
Administrative Level 1 (counties)  map of Liberia
}
\keyword{datasets}
