% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{districts}
\alias{districts}
\title{districts}
\format{
A SpatialPolygonsDataFrame with 136 features:
\describe{
\item{\code{OBJECTID}}{Object identifier}
\item{\code{CNTRY_NAME}}{Country name}
\item{\code{CNTRY_CODE}}{Three-character ISO country code. For Liberia it is
    LBR.}
\item{\code{WHO_CODE}}{Three-character WHO country code. For Liberia it is
    LIB.}
\item{\code{Shape_Leng}}{Length of shape}
\item{\code{Shape_Area}}{Area of shape}
\item{\code{admin1name}}{Name of county}
\item{\code{admin1Pcod}}{County postal code}
\item{\code{admin2name}}{Name of district}
\item{\code{admin2Pcod}}{District postal code}
\item{\code{Pcod1_iso3}}{County postal code using ISO3 code}
\item{\code{Pcod2_iso3}}{District postal code using ISO3 code}
}
}
\source{
\url{https://data.humdata.org/dataset/liberia-admin-level-2-boundaries}
}
\usage{
districts
}
\description{
Administrative Level 2 (districts)  map of Liberia
}
\keyword{datasets}
