% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{enumerationArea}
\alias{enumerationArea}
\title{enumerationArea}
\format{
A data.frame with 8 columns and 751 rows
\describe{
\item{\code{County}}{Name of county}
\item{\code{District}}{Name of district}
\item{\code{Clan}}{Name of clan}
\item{\code{EACODE}}{Enumeration area code}
\item{\code{MALE}}{Number of male population}
\item{\code{FEMALE}}{Number of female population}
\item{\code{HH}}{Number of households}
\item{\code{TOTAL}}{Total population}
}
}
\source{
Liberia Institute of Statistics and Geo-Information Services
}
\usage{
enumerationArea
}
\description{
Dataset of enumeration areas with population size for Grand Bassa and Rural
Montserrado
}
\keyword{datasets}
