% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grandBassaEA}
\alias{grandBassaEA}
\title{grandBassaEA}
\format{A SpatialPolygonsDataFrame with 468 features:
\describe{
\item{\code{CCNAME}}{County name}
\item{\code{DNAME}}{District name}
\item{\code{CLNAME}}{Clan name}
\item{\code{FEACODE}}{Full EA code}
\item{\code{SAREA}}{Area map label}
\item{\code{EFEACODE}}{Extended full EA code}
\item{\code{CCODE}}{County code}
\item{\code{DCODE}}{District code}
\item{\code{CLCODE}}{Clan code}
\item{\code{EACODE}}{Enumeration area code}
\item{\code{TOTAL}}{Total population}
\item{\code{MALE}}{Male population}
\item{\code{FEMALE}}{Female population}
\item{\code{HH}}{Total households}
\item{\code{CCODE1}}{Extended county code}
\item{\code{DCODE1}}{Extended district code}
\item{\code{CLCODE1}}{Extended clan code}
\item{\code{EACODE1}}{Extended enumeration area code}
}}
\source{
Liberia Institute of Statistics and Geo-Information Services
}
\usage{
grandBassaEA
}
\description{
Enumeration area map of Grand Bassa county
}
\keyword{datasets}
