% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{locality}
\alias{locality}
\title{locality}
\format{
A data.frame with 9 columns and 3412 rows
\describe{
\item{\code{FID}}{Locality ID}
\item{\code{County}}{Name of county}
\item{\code{District}}{Name of district}
\item{\code{Clan}}{Name of clan}
\item{\code{`Locality/ Town`}}{Name of locality/town}
\item{\code{MALE}}{Number of male population}
\item{\code{FEMALE}}{Number of female population}
\item{\code{TOTAL}}{Total population}
\item{\code{HH}}{Number of households}
}
}
\source{
Liberia Institute of Statistics and Geo-Information Services
}
\usage{
locality
}
\description{
Dataset of localities with population size for Grand Bassa and Rural
Montserrado
}
\keyword{datasets}
