% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{monrovia}
\alias{monrovia}
\title{monrovia}
\format{A data.frame with 8 columns and 751 rows
\describe{
\item{\code{County}}{Name of county}
\item{\code{District}}{Name of district}
\item{\code{Town/Locality}}{Town or locality name}
\item{\code{MALE}}{Number of male population}
\item{\code{FEMALE}}{Number of female population}
\item{\code{TOTAL}}{Total population}
}}
\source{
Liberia Institute of Statistics and Geo-Information Services
}
\usage{
monrovia
}
\description{
Dataset of enumeration areas with population size for Greater Monrovia
District
}
\keyword{datasets}
