% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{settlements}
\alias{settlements}
\title{settlements}
\format{
A SpatialPointsDataFrame with 14013 features:
\describe{
\item{\code{OBJECTID}}{Object identifier}
\item{\code{FNAME}}{Settlement name}
\item{\code{CLASS}}{A numeric value specifying the settlement class. 1 = Capital
    of country; 2 = Capital of county; 3 = Capital of district; 0 = village}
\item{\code{DESCRIPTIO}}{A string value describing the settlement class based
    on \code{CLASS}}
\item{\code{COORD_X}}{Longitude coordinate of settlement}
\item{\code{COORD_Y}}{Latitude coordiante of settlement}
\item{\code{admin1name}}{Name of county}
\item{\code{admin1Pcod}}{County postal code}
\item{\code{admin2name}}{Name of district}
\item{\code{admin2Pcod}}{District postal code}
\item{\code{admin0name}}{Name of country}
\item{\code{admin0Pcod}}{Country postal code. LR for Liberia.}
\item{\code{rowcacode0}}{Three-character ISO country code}
\item{\code{Pcod1_iso3}}{County postal code using ISO3 code}
\item{\code{Pcod2_iso3}}{District postal code using ISO3 code}
}
}
\source{
\url{https://data.humdata.org/dataset/liberia-settlements-0}
}
\usage{
settlements
}
\description{
Map of settlement locations in Liberia
}
\keyword{datasets}
